Revchart - fixed drawing error when scrolling fast
When scrolling through the story fast (especially upwards) the transitions from one step to another may get interrupted. This yields green rectangles to be drawn at erroneous heights.
I fixed it by adding the y coordinate update at all steps so that the transition is re-triggered if necessary, and the final rendering is always correct.